nat: extract source-NAT tests into tests_source.rs (#4409)
Begin the mechanical split of the 8,685-LOC nat/tests.rs (audit
refactor backlog #4409) into per-subject sibling test files loaded via
`#[cfg(test)] #[path=...] mod` from nat/mod.rs, following the #4405/#4417
validator-split and #1046 frame/tests.rs precedents.

This first step moves the 15 source (interface / address-match) NAT
tests -- interface_source_nat_*, off_rule_short_circuits_translation,
snat_all_malformed_*, snat_bare_host_*, snat_unscoped/valid/mixed, and
reverse_decision_turns_snat_into_reply_dnat -- verbatim into
tests_source.rs. No test body is edited and nothing is renamed; the
`use` header is copied verbatim and carries `#![allow(unused_imports)]`
so a per-file import subset does not warn.

nat/tests.rs retains the remaining 188 tests; it will be emptied and
removed in the final step of the split. Because the new file is a
`#[path]` child module of `nat`, `super::*` continues to resolve to the
nat module items exactly as before.
